+++ Changelog.rst
@@ -5,6 +5,8 @@ Version NEXTVERSION
 
 * Write UGRID datasets with `cfdm.write`
   (https://github.com/NCAS-CMS/cfdm/issues/271)
+* New keyword to `cfdm.read`: ``filesystem``
+  (https://github.com/NCAS-CMS/cfdm/issues/397)
 * New keyword parameter to `cfdm.Data.compute`: ``persist``
   (https://github.com/NCAS-CMS/cfdm/issues/389)
 * New function to control the persistence of computed data: